requirements.txt
asgiref==3.5.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-import-export==2.8.0
django-phone-field==1.8.1
django-tinymce==3.4.0
django-versatileimagefield==2.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.1.0
python-decouple==3.6
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
